authorization implemented

+++ src/services/notes/controllers/note-controller.js
@@ -1,19 +1,19 @@
-import { nanoid } from 'nanoid';
-import notes from '../../../utils/notes.js';
 import InvariantError from '../../../exceptions/invariant-error.js';
 import NotFoundError from '../../../exceptions/not-found-error.js';
 import response from '../../../utils/response.js';
 import NoteRepositories from '../repositories/notes-repositories.js';
+import AuthorizationError from '../../../exceptions/authorization-error.js';
 
 
 export const createNote = async (req, res, next) => {
     const { title, body, tags } = req.validated;
+    const { id: owner } = req.user;
     const note = await NoteRepositories.createNote({
         title,
         body,
-        tags
+        tags,
+        owner
     });
-
     if (!note) {
         return next(new InvariantError('Catatan gagal ditambahkan'));
     }
@@ -22,30 +22,38 @@ export const createNote = async (req, res, next) => {
 };
 
 export const getNotes = async (req, res) => {
-    const notes = await NoteRepositories.getNotes();
+    const { id: owner } = req.user;
+    const notes = await NoteRepositories.getNotes(owner);
     return response(res, 200, 'Catatan sukses ditampilkan', notes);
 };
 
 export const getNoteById = async (req, res, next) => {
     const { id } = req.params;
+    const { id: owner } = req.user;
+    const isOwner = await NoteRepositories.verifyNoteOwner(id, owner);
+    if (!isOwner) {
+        return next(new AuthorizationError('Anda tidak berhak mengakses resource ini'));
+    }
     const note = await NoteRepositories.getNoteById(id);
-
     if (!note) {
         return next(new NotFoundError('Catatan tidak ditemukan'));
     }
-
     return response(res, 200, 'Catatan sukses ditampilkan', note);
 };
 
 
-export const editNote = async (req, res, next) => {
+export const editNoteById = async (req, res, next) => {
     const { id } = req.params;
     const {
         title,
         body,
         tags
     } = req.validated;
-
+    const { id: owner } = req.user;
+    const isOwner = await NoteRepositories.verifyNoteOwner(id, owner);
+    if (!isOwner) {
+        return next(new AuthorizationError('Anda tidak berhak mengakses resource ini'));
+    }
     const note = await NoteRepositories.editNote({
         id,
         title,
@@ -55,18 +63,20 @@ export const editNote = async (req, res, next) => {
     if (!note) {
         return next(new NotFoundError('Catatan tidak ditemukan'));
     }
-
     return response(res, 200, 'Catatan berhasil diperbarui', note);
 };
 
 
-export const deleteNote = async (req, res, next) => {
+export const deleteNoteById = async (req, res, next) => {
     const { id } = req.params;
+    const { id: owner } = req.user;
+    const isOwner = await NoteRepositories.verifyNoteOwner(id, owner);
+    if (!isOwner) {
+        return next(new AuthorizationError('Anda tidak berhak mengakses resource ini'));
+    }
     const deletedNote = await NoteRepositories.deleteNote(id);
-
     if (!deletedNote) {
         return next(new NotFoundError('Catatan tidak ditemukan'));
     }
-
     return response(res, 200, 'Catatan berhasil dihapus', deletedNote);
 };
